begin test

+++ srcs/Worker.cpp
@@ -122,7 +122,7 @@ void Worker::send_client(int i)
 	size_send = write(client.fd, client.buffer_send.c_str(), size_send);
 	if (size_send == -1)
 		throw std::runtime_error(std::string(strerror(errno)));
-	if (size_send < client.buffer_send.size())
+	if ((size_t)size_send < client.buffer_send.size())
 		update_modif_list(client.fd, EVFILT_WRITE, EV_ADD | EV_ONESHOT);
 	if (!end)
 	{
